+++ src/Skills/SkillsList/SimpleAccordion/index.js
