+++ src/data/grade.js
@@ -1,0 +1,1 @@
+export const DEFAULT_GRADE = "analystDeveloper";
