+++ src/data/index.js
@@ -1,0 +1,5 @@
+import { DEFAULT_GRADE } from "./grade";
+
+export {
+  DEFAULT_GRADE
+};
